% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authenticate_ghost.R
\name{authenticate_ghost}
\alias{authenticate_ghost}
\title{authenticate_ghost}
\usage{
authenticate_ghost(username, password)
}
\arguments{
\item{username}{Ghost username (format email)}

\item{password}{Ghost password.}
}
\value{
http request content is returned. If an authentication is unsuccesful the error message is returned.
}
\description{
Make call to authenticate current user.
}
\examples{
authenticate_ghost("foo@barcom","foopassword")
}

