% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_functions_ghost.R
\name{construct_post}
\alias{construct_post}
\title{Creates a post object to be passed on when a post is created or updated.}
\usage{
construct_post(post_body)
}
\arguments{
\item{post_body}{is a vector with named objects. The title and markdown object are required. The other object are optional.
#' \itemize{
  \item title [required]: The post title (string).
  \item markdown [required] : the markdown content to be converted by Ghost to html. This contains all the content of your post (string)
  \item slug: The url where you post can be accessed. Ghost automagically create a slug based on the input. Spaces and other non alpanumeric characters will be escaped by Ghost. Any string is valid, if nothing is provided Ghost will create a slug for you. (string)
  \item status: Can either take draft or published. Default is draft. (string)
  \item featured: If a post is featured or not.Default is false (boolean).
  \item image: url to an image. Depending on the ghost template the image will be used as a cover (string)
  \item page: Indicated is the post serves as a static page. Default is false (boolean)
  \item author: The author id of the author of the post.(integer
  \item tags: list of tags objects (see ghost documentation for a description of the tag object. )
}}
}
\value{
Response object. See the \code{\link{construct_response}} function.
}
\description{
Please review the api \href{https://github.com/TryGhost/Ghost/wiki/}{unofficial documentation} from ghost to obtain official information.Please review the official documentation \href{http://api.ghost.org/v0.1-alpha/docs} once it is released.
}
\examples{
post_body <- c(
markdown = 'Content',
slug     = "Slug",
status   = "Publish",
image    = 'Image',
featured = Featured,
page     = "0",
author   = Author,
tags     = list(TagsList)
)
post_body <- construct_post(post_body)
}

