% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_functions_ghost.R
\name{construct_response}
\alias{construct_response}
\title{Creates a standarized response from an api call.}
\usage{
construct_response(ghost_response)
}
\arguments{
\item{ghost_response}{The http object returned from an api call.}
}
\value{
Response object. The response object is a named list with three objects.
#' \itemize{
  \item message: a string which can be used to provide feedback messages to the user.
  \item status: a boolean value which indicates if the api call was succesfull.
  \item content: the parsed content from the http request is the request was successful. If an error occurred the error is returned in the content.
}
}
\description{
Internal package function to create a response
}

