% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_functions_ghost.R
\name{construct_url}
\alias{construct_url}
\title{Creates the url to be passed to the api call.}
\usage{
construct_url()
}
\value{
Response object. See the \code{\link{construct_response}} function.
}
\description{
Creates the url to be passed to the api call. The current api url '/ghost/api/v0.1/' is included. the baseurl set in the RGhost environment is used to construct the url.
}
\examples{
create_post_ghost(construct_post(post_body))
}

