% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_post_ghost.R
\name{create_post_ghost}
\alias{create_post_ghost}
\title{create_post_ghost}
\usage{
create_post_ghost(post_body)
}
\arguments{
\item{post_body}{a R list object to be converted to a valid Ghost Post object. See the \code{\link{construct_body}} function for a detailed description of a valid post object.}
}
\value{
Response object. See the \code{\link{construct_response}} function.
}
\description{
Create a new post on ghost bloggin platform.
}
\examples{
create_post_ghost(construct_post(post_body))
}

