% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delete_post_ghost.R
\name{delete_post_ghost}
\alias{delete_post_ghost}
\title{delete_post_ghost}
\usage{
delete_post_ghost(id)
}
\arguments{
\item{Post}{Id from an existing Ghost post.}
}
\value{
Response object. See the \code{\link{construct_response}} function.
}
\description{
Delete a post from Ghost.
}
\examples{
delete_post_ghost(1)
}

