% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_users_ghost.R
\name{get_current_user}
\alias{get_current_user}
\title{get_current_user}
\usage{
get_current_user()
}
\value{
Response object. See the \code{\link{construct_response}} function. A single user object is returned
}
\description{
Get information on the current logged in user.
}
\examples{
get_current_user()
}

