% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authenticate_ghost.R
\name{get_ghost_token}
\alias{get_ghost_token}
\title{get_ghost_token}
\usage{
get_ghost_token()
}
\value{
Valid access token.
}
\description{
Checks if the current access token is valid and return the access_token. If the token has expired a new token is requested by calling the refresh_authenticate_ghost function.
}
\examples{
get_ghost_token()
}

