% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_posts_ghost.R
\name{get_post_ghost}
\alias{get_post_ghost}
\title{get_post_ghost}
\usage{
get_post_ghost(id)
}
\arguments{
\item{Post}{id from an existing post object.}
}
\value{
Response object. See the \code{\link{construct_response}} function. Returns a post object. Tags associated with the current blogpost are included by default.
}
\description{
Gets a signle post.
}
\examples{
get_post_ghost(1)
}

