% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_posts_ghost.R
\name{get_posts_ghost}
\alias{get_posts_ghost}
\title{Get 15 Ghosts blogposts}
\usage{
get_posts_ghost()
}
\arguments{
\item{status}{ghost posts either have status draft, published. The 'all' parameter gets both published and draft}
}
\value{
Returns latest 15 post objects in a flattened list containing post objects.
}
\description{
Get 15 Ghosts blogposts
}

