% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_tags_ghost.R
\name{get_tags}
\alias{get_tags}
\title{get_tags}
\usage{
get_tags()
}
\value{
Note a flattened list of tag objects and a nested list of tagobjects is returned. Implementation of this function to adhere to the standard in this package. The current implementation was done for convenience reasons to support the Shiny application.
}
\description{
Get all existing tags.
}
\examples{
get_tags()
}

