% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_users_ghost.R
\name{get_users}
\alias{get_users}
\title{get_users}
\usage{
get_users()
}
\value{
Response object. See the \code{\link{construct_response}} function. An list of user objects is returned.
}
\description{
Get all users associated with the current logged in account.
}
\examples{
get_users()
}

