% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authenticate_ghost.R
\docType{data}
\name{ghost}
\alias{ghost}
\title{set_credentials}
\format{\preformatted{<environment: 0x7fb183633470> 
}}
\usage{
ghost
}
\arguments{
\item{username}{Ghost username (format email)}

\item{password}{Ghost password.}

\item{url}{Ghost Url (using https:// format).}
}
\value{
Access token required for any request to the url \code{access_token} and \code{y}.
}
\description{
set initial variables to be used in the package and authenticates the user to obtain an access token
}
\examples{
set_credentials("foo@barcom","foopassword","http://foo.com")
}
\keyword{datasets}

