% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_launcher_rghost.R
\name{launch_rghost_app}
\alias{launch_rghost_app}
\title{launch_rghost_app}
\usage{
launch_rghost_app()
}
\description{
Launches the associated shiny application. The shiny application implements the RGhost package and allows the user to take advantage of parsing R code to markdown and then posting this to Ghost. See the file for more information.
}
\examples{
run_RGhost_app ()
}

