% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_functions_ghost.R
\name{parse_content}
\alias{parse_content}
\title{Converts the return JSON object into a nested R list.}
\usage{
parse_content(post_body)
}
\arguments{
\item{post_body}{a JSON object containing the content to be converted to R.}
}
\value{
Nested r list representation of the parsed object.
}
\description{
Converts the return JSON object into a nested R list. Currently using jsonlite to parse the content.
}

