% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authenticate_ghost.R
\name{refresh_authenticate_ghost}
\alias{refresh_authenticate_ghost}
\title{refresh_authenticate_ghost}
\usage{
refresh_authenticate_ghost(refresh_token)
}
\arguments{
\item{refresh}{token.}
}
\value{
New access token which is valid for the next hour.
}
\description{
Gets a new access token if the token is expired. Ghost access tokens expire after one hour. If authentication is succesful an refresh token is returned in the authetication call.
}
\examples{
refresh_authenticate_ghost("refresh_token")
}

