% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGhost-package.R
\docType{package}
\name{RGhost}
\alias{RGhost}
\alias{RGhost-package}
\title{RGhost: A R wrapper of the Ghost blogging platform api.}
\description{
The package supports most Ghost functions in order to create, update and delete posts.
Currently the Ghost Api is under development and not officially released to the public, as such some
The package can be implemented in R or a Shiny application as an application in the package.
}
\section{RGhost functions}{

The RGhost function are the following:
\itemize{
  \item Authenticate
  \item Create post
  \item Get a list of posts, get a specific post
  \item Get information on the current user
  \item Get a list of users
  \item Get a list of tags currently used.
  \item Update and existing post
  \item Delete a specific post
}
}

