% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_post_ghost.R
\name{update_post_ghost}
\alias{update_post_ghost}
\title{update_post_ghost}
\usage{
update_post_ghost(id, post_body)
}
\arguments{
\item{id}{Post id of an existing post.}

\item{post_body}{a valid post object.}
}
\value{
Response object. See the \code{\link{construct_response}} function.
}
\description{
Update and existing post. Tags are included by default.
}
\examples{
update_post_ghost(1,construct_body(post_body))
}

