% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authenticate_ghost.R
\name{authenticate_ghost}
\alias{authenticate_ghost}
\title{Authenticate ghost serive}
\usage{
authenticate_ghost()
}
\arguments{
\item{username}{Ghost username (format email)}

\item{password}{Ghost password.}

\item{url}{Ghost Url (using https:// format).}
}
\value{
Access token required for any request to the url \code{access_token} and \code{y}.
}
\description{
Authenticate ghost serive
}

