% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_posts_ghost.R
\name{get_post_ghost}
\alias{get_post_ghost}
\title{Get single post}
\usage{
get_post_ghost(id)
}
\arguments{
\item{id}{postid to fetch}
}
\value{
Returns latest 1 post objects.
}
\description{
Get single post
}

