% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_posts_ghost.R
\name{get_posts_ghost}
\alias{get_posts_ghost}
\title{get_posts_ghost}
\usage{
get_posts_ghost()
}
\value{
Response object. See the \code{\link{construct_response}} function. An list of post objects is returned.
}
\description{
Get all existing posts. Posts, static pages, draft and published posts are currently returned. Future implementation will allow to pass parameters to this function to select the returned posts.
}
\examples{
get_posts_ghost()
}

