% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authenticate_ghost.R
\docType{data}
\name{ghost}
\alias{ghost}
\title{Base input parameters}
\format{\preformatted{<environment: 0x7fc7fc40ec28> 
}}
\usage{
ghost
}
\arguments{
\item{username}{Ghost username (format email)}

\item{password}{Ghost password.}

\item{url}{Ghost Url (using https:// format).}
}
\value{
Access token required for any request to the url \code{access_token} and \code{y}.
}
\description{
Base input parameters
}
\keyword{datasets}

